dom/html/test/test_bug569955.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=569955
-->
<head>
  <title>Test for Bug 569955</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=569955">Mozilla Bug 569955</a>
<p id="display"></p>
<div id="content" style="display: none">
  <input id='i' autofocus>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 569955 **/

function runTests()
{
  isnot(document.activeElement, document.getElementById('i'),
        "not rendered elements can't be autofocused");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  setTimeout(runTests, 0);
});

</script>
</pre>
</body>
</html>
